Revert "test de double : pb de taille!"
This reverts commit 8e690dff0fc4a85cf8d99ae04bda8dc8399a4128.

+++ b_hardware.cpp
@@ -114,9 +114,9 @@
 #define PIN_LOW_POWER 1
 
 // prototype
-inline double _fclamp(double, double, double);
-inline double _fmin(double, double);
-inline double _fmax(double, double);
+inline float _fclamp(float, float, float);
+inline float _fmin(float, float);
+inline float _fmax(float, float);
 
 using namespace daisy;
